+++ R/utils.R
@@ -3,11 +3,11 @@ globalVariables(".")
 
 # Algorithm functions and classes
 ALG.NAME <- c("pam", "svm", "rf", "lda", "slda", "sdda", "mlr_glm", "mlr_lasso",
-              "mlr_ridge", "mlr_enet", "mlr_nnet", "nnet", "nbayes", "adaboost",
+              "mlr_ridge", "mlr_enet", "mlr_nnet", "nnet", "nbayes",
               "adaboost_m1", "xgboost", "knn")
 ALG.CLASS <- c("pamrtrained", "train", "svm", "randomForest", "lda", "sda",
                "cv.glmnet", "glmnet", "multinom", "nnet.formula", "naiveBayes",
-               "maboost", "boosting", "xgb.Booster", "knn")
+               "boosting", "xgb.Booster", "knn")
 
 # Algorithms that need all continuous predictors
 ALG.CONT <- c("svm", "lda", "mlr_glm", "mlr_lasso", "mlr_ridge", "mlr_enet")